simplify sending of [turn] data
1) We remove the syncmp_handler class which was a
   bit too complicated for its rather simple task
2) We move network related stuff out of playsingle_controller
