+++ src/syncmp_handler.cpp
@@ -14,53 +14,3 @@
 */
 
 #include "syncmp_handler.hpp"
-
-#include <cassert>
-#include <algorithm>
-
-syncmp_handler::syncmp_handler()
-{
-	syncmp_registry::add_handler(this);
-}
-
-syncmp_handler::~syncmp_handler()
-{
-	syncmp_registry::remove_handler(this);
-}
-
-std::vector<syncmp_handler*>& syncmp_registry::handlers()
-{
-	//using pointer in order to prevent destruction at program end. Although in this simple case it shouldn't matter.
-	static handler_list* handlers_ = new handler_list();
-	return *handlers_;
-}
-
-void syncmp_registry::remove_handler(syncmp_handler* handler)
-{
-	handler_list::iterator elem = std::find(handlers().begin(), handlers().end(), handler);
-	assert(elem != handlers().end());
-	handlers().erase(elem);
-}
-
-void syncmp_registry::add_handler(syncmp_handler* handler)
-{
-	handler_list::iterator elem = std::find(handlers().begin(), handlers().end(), handler);
-	assert(elem == handlers().end());
-	handlers().push_back(handler);
-}
-
-void syncmp_registry::pull_remote_choice()
-{
-	for(syncmp_handler* phandler : handlers())
-	{
-		phandler->pull_remote_choice();
-	}
-}
-
-void syncmp_registry::send_user_choice()
-{
-	for(syncmp_handler* phandler : handlers())
-	{
-		phandler->send_user_choice();
-	}
-}
